Solve 07_tempConversion
Modified the convertToCelsius function such that it takes a fahrenheit
temperature and returns the celsius temperature rounded to one decimal
place. Modified the convertToFahrenheit function such that it takes a
celsius tempertaure and returns the fahrenheit temperature rounded to
one decimal place. This implementation passes all tests outlined in
tempConversion.spec.js.