Solve 12_findTheOldest
Created the findTheOldest function, which sorts an array of people
objects by age and returns the oldest person. The findTheOldestReduce
function was also created, which uses the reduce method instead of the
sort method. This second function was created after viewing the exercise
solution in order to enhance understanding of a more elegant solution.
Both implementations pass the tests outlined in findTheOldest.spec.js.