Add 09_palindromes regex solution
Created the palindromesRegex function, which uses regex to remove all
non-alphanumeric characters from a string and then determines if the
reversed string is the same as the forward string (i.e., is the string
a palindrome). This alternative function was created after viewing the
exercise solution in order to enhance understanding of a more elegant
solution. This implementation also passes the tests outlined in
palindromes.spec.js.